refactor: LocationHelper 정의
