+++ Diary/DiaryContentView/Helper/LocationHelper.swift
@@ -1,0 +1,42 @@
+//
+//  LocationHelper.swift
+//  Diary
+//
+//  Created by Rowan, Harry on 2023/05/11.
+//
+
+import CoreLocation
+
+final class LocationHelper: NSObject, CLLocationManagerDelegate {
+    private let locationManager = CLLocationManager()
+    var eventHandler: ((CLLocationCoordinate2D) -> Void)?
+    
+    override init() {
+        super.init()
+        
+        locationManager.delegate = self
+    }
+    
+    func setEventHandler(_ closure: @escaping (CLLocationCoordinate2D) -> Void) {
+        self.eventHandler = closure
+    }
+    
+    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
+        switch manager.authorizationStatus {
+        case .authorizedWhenInUse, .authorizedAlways:
+            locationManager.startUpdatingLocation()
+        case .restricted, .notDetermined:
+            locationManager.requestWhenInUseAuthorization()
+        case .denied:
+            locationManager.requestWhenInUseAuthorization()
+        default:
+            break
+        }
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        guard let coordinate = locations.last?.coordinate else { return }
+
+        eventHandler?(coordinate)
+    }
+}
